+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @ni/nimble-components
 
-<!-- This log was last generated on Sat, 11 Oct 2025 04:58:50 GMT and should not be manually modified. -->
+<!-- This log was last generated on Fri, 31 Oct 2025 20:33:25 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 34.5.0
+
+Fri, 31 Oct 2025 20:33:25 GMT
+
+### Minor changes
+
+- Add new icons for camera, scanner gun, DIAdem, calipers ([ni/nimble@7bc6318](https://github.com/ni/nimble/commit/7bc6318cc1a2984d4411936acc6c9adbe0a3c65b))
+- Bump @ni/nimble-tokens to v8.13.0
+
+## 34.4.0
+
+Tue, 28 Oct 2025 15:26:27 GMT
+
+### Minor changes
+
+- Adding Nimble Chip component ([ni/nimble@6164b3e](https://github.com/ni/nimble/commit/6164b3ea08f56709d45ed471914f523618c226bf))
 
 ## 34.3.3