+++ packages/ok-components/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @ni/ok-components
 
-<!-- This log was last generated on Sat, 11 Oct 2025 04:58:50 GMT and should not be manually modified. -->
+<!-- This log was last generated on Fri, 31 Oct 2025 20:33:25 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.1.0
+
+Fri, 31 Oct 2025 20:33:25 GMT
+
+### Minor changes
+
+- Bump @ni/nimble-components to v34.5.0
+- Bump @ni/spright-components to v6.4.0
+
+## 0.0.6
+
+Tue, 28 Oct 2025 15:26:27 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v34.4.0
+- Bump @ni/spright-components to v6.3.4
 
 ## 0.0.5