+++ packages/react-workspace/nimble-react/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @ni/nimble-react
 
-<!-- This log was last generated on Sat, 11 Oct 2025 22:38:51 GMT and should not be manually modified. -->
+<!-- This log was last generated on Fri, 31 Oct 2025 20:33:25 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.5.0
+
+Fri, 31 Oct 2025 20:33:25 GMT
+
+### Minor changes
+
+- Bump @ni/nimble-tokens to v8.13.0
+- Bump @ni/nimble-components to v34.5.0
+
+## 0.4.1
+
+Tue, 28 Oct 2025 15:26:27 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v34.4.0
 
 ## 0.4.0